Adding TestUtil. Working on Insurance and blackjack tests still.

+++ include/Bet.h
@@ -8,18 +8,18 @@ class Bet
     friend class Sim;
     friend class Player;
 public:
-    inline int    GetAmount(){ return _amount; }
+    inline double    GetAmount(){ return _amount; }
     inline double GetPayout(){ return _payout; }
     Bet(){}
-    Bet(int amount) : _amount(amount) {}
-    Bet(int amount, double payout);
+    Bet(double amount) : _amount(amount) {}
+    Bet(double amount, double payout);
     ~Bet()
     {
         //std::cout << "Bet dtor." << std::endl;
     }
 
 private:
-    int _amount;
+    double _amount;
     double _payout;
 
 };
